utlity is moved from generic to my module
