+++ Facebook/src/test/java/TestFB.java
@@ -2,7 +2,6 @@ import Pages.Signup;
 import base.CommonAPI;
 import org.testng.Assert;
 import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.Parameters;
 import org.testng.annotations.Test;
 
 public class TestFB extends CommonAPI {
