+++ pom.xml
@@ -11,7 +11,6 @@
     <modules>
         <module>generic</module>
         <module>Facebook</module>
-        <module>nbc</module>
     </modules>
     <dependencyManagement>
         <dependencies>
